Add a base class for visibility handlers and use it in PreferencesHandler
